Add files via upload
I translate documents from Japanese to English as a side job. Every month I have to create an invoice and email to send back charging for my services. It was a hassle to fill in the dates, charges etc. so I created this program to automate the process. It looks at the current date and determine the appropriate dates to fill into the respective cells in the excel sheet. It also adds the fees in the appropriate month, saves them to the payments files and adds the total amount to the invoice. The word document is also simply updated to the correct months. What use to require 10 minutes of focus, now can be completed with a double mouse-click.